fix unnecessary commits when branch does not change

+++ src/pootle-api/to_liferay.sh
@@ -50,20 +50,27 @@ function do_commit() {
 		base_src_dir=${PATH_BASE_DIR[$i]}
 		cd $base_src_dir
 		logt 2 "$base_src_dir"
-		if exists_branch "pootle_export" "$base_src_dir"; then
-		    logt 3 -n "Cleaning old export branch: git branch -D pootle_export"
-            git branch -D pootle_export > /dev/null 2>&1
+		if something_changed; then
+            if exists_branch "pootle_export" "$base_src_dir"; then
+                logt 3 -n "Cleaning old export branch: git branch -D pootle_export"
+                git branch -D pootle_export > /dev/null 2>&1
+                check_command
+            fi
+            logt 3 -n "Creating new export branch: git checkout -b pootle_export"
+            git checkout -b pootle_export > /dev/null 2>&1
             check_command
+            msg="Pootle export, created by $product"
+            logt 3 -n "git commit -m $msg"
+            git commit -m "\"$msg\""
+            check_command
+            logt 3 -n "git push origin pootle_export"
+            git push -f origin pootle_export > /dev/null 2>&1
+            check_command
+		else
+		    logt 3 "No changes to commit!!"
 		fi
-		logt 3 -n "Creating new export branch: git checkout -b pootle_export"
-		git checkout -b pootle_export > /dev/null 2>&1
-		check_command
-		msg="Pootle export, created by $product"
-		logt 3 -n "git commit -a -m $msg"
-		git commit -a -m "$msg"  > /dev/null 2>&1
-		check_command
-		logt 3 -n "git push origin pootle_export"
-		git push origin pootle_export > /dev/null 2>&1
+		logt 3 -n "git checkout master"
+		git checkout master > /dev/null 2>&1
 		check_command
 	done;
 	cd $old_dir
@@ -78,11 +85,11 @@ function update_pootle_files() {
 	do
 		project=${PROJECT_NAMES[$i]}
 		logt 2 "  $project"
-		logt 3 -n "    Synchronizing pootle stores for all languages "
+		logt 3 -n "Synchronizing pootle stores for all languages "
 		# Save all translations currently in database to the file system
 		$POOTLEDIR/manage.py sync_stores --project="$project" -v 0 > /dev/null 2>&1
 		check_command
-		logt 3 "    Copying exported tranlsations into working dir"
+		logt 3 "Copying exported tranlsations into working dir"
 		for language in $(ls "$PODIR/$project"); do
 		    logt 0 -n  "$(get_locale_from_file_name $language) "
     		cp -f  "$PODIR/$project/$language" "$TMP_PROP_OUT_DIR/$project/"
